+++ test/test_add_color.py
@@ -4,7 +4,6 @@ import shutil
 import ctview.add_color as add_color
 from ctview.utils_folder import dico_folder_template
 
-TMP = "tmp"
 COORDX = 77055
 COORDY = 627760
 TILE_COORD_SCALE = 10
@@ -19,7 +18,8 @@ INPUT_DIR_RASTER = os.path.join("data", "raster")
 INPUT_LAZ_TILENAME_WITHOUT_COLOR = f"test_data_{COORDX}_{COORDY}_LA93_IGN69.laz"
 INPUT_RASTER_WITHOUT_COLOR = os.path.join(INPUT_DIR_RASTER, f"test_data_{COORDX}_{COORDY}_LA93_IGN69_interp.tif")
 
-OUTPUT_DIR_COLOR = os.path.join("tmp", "color")
+OUTPUT_DIR = os.path.join("tmp", "add_color")
+OUTPUT_DIR_COLOR = os.path.join(OUTPUT_DIR, "color")
 OUTPUT_DIR_LUT = os.path.join(OUTPUT_DIR_COLOR, "LUT")
 
 EXPECTED_DTM_COLOR_1CYCLE = os.path.join(
@@ -32,13 +32,13 @@ EXPECTED_DTM_COLOR_5CYCLES = os.path.join(
 
 def setup_module(module):
     try:
-        shutil.rmtree(TMP)
+        shutil.rmtree(OUTPUT_DIR)
 
     except FileNotFoundError:
         pass
-    os.mkdir(TMP)
-    os.mkdir(OUTPUT_DIR_COLOR)
-    os.mkdir(OUTPUT_DIR_LUT)
+    os.makedirs(OUTPUT_DIR)
+    os.makedirs(OUTPUT_DIR_COLOR)
+    os.makedirs(OUTPUT_DIR_LUT)
 
 
 def test_color_raster_dtm_hillshade_with_LUT():