+++ test/test_clip_raster.py
@@ -1,4 +1,5 @@
 import os
+import shutil
 
 from ctview.clip_raster import crop_raster
 from ctview.utils_gdal import get_raster_corner_coord, transform_CornerCoord_to_Bounds
@@ -8,7 +9,18 @@ DIR_LAS = os.path.join("data", "las")
 DIR_RASTER = os.path.join("data", "raster")
 IN_LAS = os.path.join(DIR_LAS, "test_data_0000_0000_LA93_IGN69_ground.las")
 IN_RASTER = os.path.join(DIR_RASTER, "test_data_0000_0000_LA93_IGN69_ground.tif")
-OUT_RASTER = os.path.join(DIR_RASTER, "test_data_0000_0000_LA93_IGN69_ground_crop.tif")
+OUTPUT_DIR = os.path.join("tmp", "clip_raster")
+OUT_RASTER = os.path.join(OUTPUT_DIR, "test_data_0000_0000_LA93_IGN69_ground_crop.tif")
+
+
+def setup_module(module):
+    try:
+        shutil.rmtree(OUTPUT_DIR)
+
+    except FileNotFoundError:
+        pass
+    os.makedirs(OUTPUT_DIR)
+
 
 ([xmin_las, xmax_las], [ymin_las, ymax_las]) = get_bounds_from_las(in_las=IN_LAS)